plugin-EnergyCampus: outsource threadworker into lib
